tools/lib/lockdep: Define pr_cont()
lockdep.c now also uses pr_cont(), so we need to implement it in
liblockdep.

It is currently always used to continue warning lines, so define
pr_cont() the same as pr_warn().  If this changes, we might need to
record the last log level in a TLS variable and have pr_cont() check
that.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name tools-lib-lockdep-define-pr_cont.patch