ALSA: timer: Avoid ABI change in 4.13.13
Commit 9b7d869ee5a7 ("ALSA: timer: Limit max instances per timer")
added two new members to the end of struct snd_timer, and this changes
ABI for many exported symbols.  However, snd_timer is always allocated
in the snd-timer module, which is the same place the new members are
used.  We can safely hide the change from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name alsa-timer-avoid-abi-change-in-4.13.13.patch