sched: Avoid ABI change in 3.2.94
The backport of commits e0e5070b20e0 "sched: add macros to define bitops
for task atomic flags" added a new field to struct task_struct which
will change ABI for most symbols in the kernel.

However, task_struct is always allocated by built-in code and the
new field is only accessed by built-in code.  So move it to the end and
hide it from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name sched-avoid-abi-change-in-3.2.94.patch