ip_fib: Avoid ABI change in 3.16.50
Commit 3fb07daff8e9 "ipv4: add reference counting to metrics" changed
the type of fib_info::fib_metrics from u32 * to dst_metrics *.
However, dst_metrics begins with the same array-of-u32 as before, so
the change doesn't break existing readers; and all writers to the
field are built-in.  So hide the change from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name ip_fib-avoid-abi-change-in-3.16.50.patch