dm: Avoid ABI change in 3.2.41
Commit fd7c092e711e 'dm: fix truncated status strings', backported
into 3.2.41, changed the return type of target_type::status from void
to int.  But we can just as easily leave it as int and ignore the
returned value.

Also, genksyms is too stupid to understand that 'unsigned' is the
same thing as 'unsigned int'.


Gbp-Pq: Topic debian
Gbp-Pq: Name dm-avoid-ABI-change-in-3.2.41.patch