iwlwifi: Do not request unreleased firmware
The iwlwifi driver currently supports firmware API versions 4-6 for
IWL6000/6300 devices.  It will request the file for the latest
supported version and then fall back to earlier versions.  However,
the latest version that has actually been released is 4, so we expect
the requests for versions 6 and then 5 to fail.

Similarly for IWL6005/6205 devices it requests API version 6, which is
available but wasn't released until long after this version of the
driver (and after the wheezy freeze).

The installer appears to report any failed request, and it is probably
not easy to detect that this particular failure is harmless.  So stop
requesting the unreleased firmware.


Gbp-Pq: Topic debian
Gbp-Pq: Name iwlwifi-do-not-request-unreleased-firmware.patch