fs: Avoid ABI change in 3.16.57
The backport of commit 95dd77580ccd "fs: Teach path_connected to
handle nfs filesystems with multiple roots." adds the field
super_block::s_iflags.

struct super_block is always allocated by the core kernel
(alloc_super() function) and not embedded in other structures,
so we can safely add fields at the end.  Move the new field
there and hide it from genksyms.

The field still needs to be set by the nfs module, so we need to
prevent loading the updated nfs module on an old kernel which doesn't
allocate enough memory for that field.  To ensure that, add a trivial
wrapper around sget() and make nfs call that.


Gbp-Pq: Topic debian
Gbp-Pq: Name fs-avoid-abi-change-in-3.16.57.patch