ARM: dts: add Raspberry Pi Compute Module 3 and IO board
The Raspberry Pi Compute Module 3 (CM3) and the Raspberry Pi
Compute Module 3 Lite (CM3L) are SoMs which contains a BCM2837 processor,
1 GB RAM and a GPIO expander. The CM3 has a 4 GB eMMC, but on the CM3L
the eMMC is unpopulated and it's up to the user to connect their
own SD/MMC device. The dtsi file is designed to work for both modules.
There is also a matching carrier board which is called
Compute Module IO Board V3.

Signed-off-by: Stefan Wahren <stefan.wahren@i2se.com>

Gbp-Pq: Topic features/arm
Gbp-Pq: Name ARM-dts-add-Raspberry-Pi-Compute-Module-3-and-IO-boa.patch