tcp: Avoid ABI change for DoS fixes
"tcp: tcp_fragment() should apply sane memory limits" adds a new Linux
MIB counter.  This adds another element to the array in struct
linux_mib.  Since that is always allocated by built-in code, it's a
backward-compatible change and we can hide the added element from
genksyms.

"tcp: add tcp_min_snd_mss sysctl" adds a new per-netns sysctl and a
new members in struct netns_ipv4.  Since this is embedded in struct
net, it changes the offsets of all the following members.  However
struct net itself is not embedded in anything, and is always allocated
by built-in code.  So move the new member to the end of struct net,
and hide it from genksyms.

Also hide the added element and member from modules, as they won't be
able to rely on their being present until we bump ABI.


Gbp-Pq: Topic debian/abi
Gbp-Pq: Name tcp-avoid-abi-change-for-dos-fixes.patch